Rename StandardDeployJob --> DeployJob (#1791)
