+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.test/META-INF/MANIFEST.MF
@@ -9,8 +9,7 @@ Bundle-Version: 0.1.0.qualifier
 Fragment-Host: com.google.cloud.tools.eclipse.appengine.deploy
 Require-Bundle: org.hamcrest;bundle-version="1.1.0",
  org.junit;bundle-version="4.12.0"
-Import-Package: com.google.cloud.tools.eclipse.appengine.deploy.standard,
- com.google.cloud.tools.eclipse.appengine.facets,
+Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.test.util,
  com.google.cloud.tools.eclipse.test.util.project,
  org.eclipse.core.expressions,
