+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/META-INF/MANIFEST.MF
@@ -14,7 +14,6 @@ Require-Bundle: com.google.cloud.tools.appengine;bundle-version="0.2.6",
 Import-Package: com.google.api.client.auth.oauth2;version="[1.22.0,1.23.0)",
  com.google.cloud.tools.eclipse.appengine.deploy,
  com.google.cloud.tools.eclipse.appengine.deploy.flex,
- com.google.cloud.tools.eclipse.appengine.deploy.standard,
  com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.appengine.ui,
  com.google.cloud.tools.eclipse.googleapis,
