+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/src/com/google/cloud/tools/eclipse/appengine/deploy/ui/DeployCommandHandler.java
@@ -19,9 +19,9 @@ package com.google.cloud.tools.eclipse.appengine.deploy.ui;
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.cloud.tools.appengine.api.deploy.DefaultDeployConfiguration;
 import com.google.cloud.tools.eclipse.appengine.deploy.CleanupOldDeploysJob;
+import com.google.cloud.tools.eclipse.appengine.deploy.DeployJob;
 import com.google.cloud.tools.eclipse.appengine.deploy.DeployPreferences;
 import com.google.cloud.tools.eclipse.appengine.deploy.DeployPreferencesConverter;
-import com.google.cloud.tools.eclipse.appengine.deploy.standard.StandardDeployJob;
 import com.google.cloud.tools.eclipse.googleapis.IGoogleApiFactory;
 import com.google.cloud.tools.eclipse.login.IGoogleLoginService;
 import com.google.cloud.tools.eclipse.sdk.ui.MessageConsoleWriterOutputLineListener;
@@ -122,11 +122,10 @@ public class DeployCommandHandler extends AbstractHandler {
     boolean includeOptionalConfigurationFiles =
         new DeployPreferences(project).isIncludeOptionalConfigurationFiles();
 
-    StandardDeployJob deploy =
-        new StandardDeployJob(project, credential, workDirectory,
-                              new MessageConsoleWriterOutputLineListener(outputStream),
-                              new MessageConsoleWriterOutputLineListener(outputStream),
-                              deployConfiguration, includeOptionalConfigurationFiles);
+    DeployJob deploy = new DeployJob(project, credential, workDirectory,
+        new MessageConsoleWriterOutputLineListener(outputStream),
+        new MessageConsoleWriterOutputLineListener(outputStream),
+        deployConfiguration, includeOptionalConfigurationFiles);
     messageConsole.setJob(deploy);
     deploy.addJobChangeListener(new JobChangeAdapter() {
