+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/DeployJob.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.eclipse.appengine.deploy.standard;
+package com.google.cloud.tools.eclipse.appengine.deploy;
 
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.cloud.tools.appengine.api.deploy.DefaultDeployConfiguration;
@@ -23,11 +23,6 @@ import com.google.cloud.tools.appengine.cloudsdk.process.ProcessExitListener;
 import com.google.cloud.tools.appengine.cloudsdk.process.ProcessOutputLineListener;
 import com.google.cloud.tools.appengine.cloudsdk.process.ProcessStartListener;
 import com.google.cloud.tools.appengine.cloudsdk.process.StringBuilderProcessOutputLineListener;
-import com.google.cloud.tools.eclipse.appengine.deploy.AppEngineDeployOutput;
-import com.google.cloud.tools.eclipse.appengine.deploy.AppEngineProjectDeployer;
-import com.google.cloud.tools.eclipse.appengine.deploy.DeployStaging;
-import com.google.cloud.tools.eclipse.appengine.deploy.Messages;
-import com.google.cloud.tools.eclipse.appengine.deploy.WarPublisher;
 import com.google.cloud.tools.eclipse.login.CredentialHelper;
 import com.google.cloud.tools.eclipse.sdk.CollectingLineListener;
 import com.google.cloud.tools.eclipse.ui.util.WorkbenchUtil;
@@ -63,7 +58,7 @@ import org.eclipse.core.runtime.SubMonitor;
  * It uses a work directory where it will create separate directories for the exploded WAR and the
  * staging results.
  */
-public class StandardDeployJob extends WorkspaceJob {
+public class DeployJob extends WorkspaceJob {
 
   private static final String STAGING_DIRECTORY_NAME = "staging";
   private static final String EXPLODED_WAR_DIRECTORY_NAME = "exploded-war";
@@ -85,13 +80,25 @@ public class StandardDeployJob extends WorkspaceJob {
   private final boolean includeOptionalConfigurationFiles;
   private final CollectingLineListener errorCollectingLineListener;
 
-  public StandardDeployJob(IProject project,
-                           Credential credential,
-                           IPath workDirectory,
-                           ProcessOutputLineListener stagingStdoutLineListener,
-                           ProcessOutputLineListener stderrLineListener,
-                           DefaultDeployConfiguration deployConfiguration,
-                           boolean includeOptionalConfigurationFiles) {
+  /**
+   * @param workDirectory temporary work directory the job can safely use (e.g., for creating and
+   *     copying various files to stage and deploy)
+   * @param stagingStdoutLineListener {@link ProcessOutputLineListener} passed to {@link CloudSdk}
+   *     to capture the staging operation stdout (where {@code appcfg.sh} outputs user-visible
+   *     log messages)
+   * @param stderrLineListener {@link ProcessOutputLineListener} passed to {@link CloudSdk} to
+   *     capture the deploy operation stderr (where {@code gcloud app deploy} outputs user-visible
+   *     log messages)
+   * @param deployConfiguration configuration passed to {@link CloudSdk} that describes what and
+   *     how to deploy
+   * @param includeOptionalConfigurationFiles if true, deploys optional XML configuration files
+   *     (e.g., {@code queue.yaml}) together
+   */
+  public DeployJob(IProject project, Credential credential, IPath workDirectory,
+      ProcessOutputLineListener stagingStdoutLineListener,
+      ProcessOutputLineListener stderrLineListener,
+      DefaultDeployConfiguration deployConfiguration,
+      boolean includeOptionalConfigurationFiles) {
     super(Messages.getString("deploy.standard.runnable.name")); //$NON-NLS-1$
     this.project = project;
     this.credential = credential;
@@ -244,7 +251,7 @@ public class StandardDeployJob extends WorkspaceJob {
   }
 
   @VisibleForTesting
-  public static String getDeployedAppUrl(boolean promoted, AppEngineDeployOutput deployOutput) {
+  static String getDeployedAppUrl(boolean promoted, AppEngineDeployOutput deployOutput) {
     String version = deployOutput.getVersion();
     String service = deployOutput.getService();
     String projectId = deployOutput.getProject();
